requirements.txt
requests
flask
beautifulsoup4
googlesearch-python
regex